ci: test auto-publish pipeline
