+++ .changeset/fair-islands-float.md
@@ -1,0 +1,5 @@
+---
+"pnpm-lock-goto": patch
+---
+
+test pipeline
