+++ .github/workflows/release-routine.yml
@@ -1,9 +1,11 @@
 # create pr with bumpped version, if merged auto-release
 name: Release-routine
 on:
-  push:
+  pull_request:
     branches:
-      - main
+      - 'main'
+    types:
+      - closed
 
 concurrency: ${{ github.workflow }}-${{ github.ref }}
 
@@ -13,7 +15,7 @@ permissions:
 
 jobs:
   release:
-    if: ${{ !startsWith(github.head_ref, 'changeset') && !startsWith(github.head_ref, 'release') }}
+    if: ${{ github.event.pull_request.merged == true && !startsWith(github.head_ref, 'changeset') && !startsWith(github.head_ref, 'release') }}
     name: Release-routine
     runs-on: ubuntu-latest
     steps:
